+++ CHANGES.rst
@@ -1,6 +1,21 @@
 Changelog
 =========
 
+1.0b1 (unreleased)
+------------------
+
+- Allow to re-index all files
+  [ebrehault]
+
+- Support different field name for the file content, and allow to map metadata fields
+  [ebrehault]
+
+- Compliancy with Plone 5.1
+  [ebrehault]
+
+- Allow to re-index all files
+  [ebrehault]
+
 
 1.0a3 (2022-10-10)
 ------------------